1-string_nconcat

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -13,8 +13,8 @@ char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
 	char *str;
 
-	size_t len1 = 0;
-	size_t len2 = 0;
+	unsigned int len1 = 0;
+	unsigned int len2 = 0;
 	unsigned int i = 0;
 
 	if (s1 == NULL)
